upgrade to new title resolver
